1. 개인정보 탐지 Rest > grpc 변경 2. 가드레일 로그 추가

+++ src/main/java/com/xcurenet/logvault/tool/cli/sample/InternalDataSampler.java
@@ -142,7 +142,9 @@ public class InternalDataSampler implements Callable<Integer> {
 
 		if (doc.getJSONObject("http").get("header") != null) {
 			sb.append("HDRFILE : ").append(header).append("\n");
-			Files.writeString(Paths.get(getPath(header)), doc.getJSONObject("http").getString("header"), StandardOpenOption.CREATE, StandardOpenOption.TRUNCATE_EXISTING);
+			if(Files.exists(Paths.get(getPath(header)))) {
+				Files.writeString(Paths.get(getPath(header)), doc.getJSONObject("http").getString("header"), StandardOpenOption.CREATE, StandardOpenOption.TRUNCATE_EXISTING);
+			}
 		}
 
 		if (doc.getJSONObject("body").getInteger("size") > 0) {
